+++ packages/bdsg/src/color-utils.ts
@@ -2,23 +2,8 @@
  * Color space conversion utilities
  */
 
-/**
- * RGB color (0-255)
- */
-export interface RGB {
-	r: number;
-	g: number;
-	b: number;
-}
-
-/**
- * HSL color (h: 0-360, s: 0-100, l: 0-100)
- */
-export interface HSL {
-	h: number;
-	s: number;
-	l: number;
-}
+export type { RGB, HSL } from "./types/color-utils.types";
+import type { RGB, HSL } from "./types/color-utils.types";
 
 /**
  * Normalize hex color string